refactor(mobile): reorganize timeline selector components and imports
